fix(lightbox): update lightbox open button after portfolio sort
